# src/main/resources/openapi/post.yml
openapi: 3.0.0
info:
  version: 1.0.2
  title: Post Service
  description: Service for posting articles about different programming languages.
  contact:
    name: Daria Smirnova
    email: smirnova@reckue.com
    url: https://github.com/daria80
servers:
  - url: https://reckue.com/api/v1

# Paths
paths:
...
  /tags:
    get:
      summary: List all tags.
      operationId: getAllTags
      tags:
        - Tags
      parameters:
        - name: limit
          in: query
          description: Limit of tags on page. By default, 10.
          required: false
          example: 10
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          description: Offset tags on page. By default, 0.
          required: false
          example: 0
          schema:
            type: integer
            format: int32
        - name: sort
          in: query
          description: Sorting tags by parameter. By default, sorts by id.
          required: false
          example: id
          schema:
            type: string
        - name: desc
          in: query
          description: Flag to sorting tags by descending. By default, false.
          required: false
          example: false
          schema:
            type: boolean
      responses:
        200:
          $ref: '#/components/responses/TagListResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/TagNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a new tag.
      operationId: createTag
      tags:
        - Tags
      security:
        - bearerAuth: []
      requestBody:
        description: Body of tag request to create a new tag.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
            example:
              name: My tag
      responses:
        201:
          $ref: '#/components/responses/TagResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/TagNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update the tag by id.
      operationId: updateTag
      tags:
        - Tags
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Body of tag request to update the tag by id.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
            example:
              name: My tag
      responses:
        201:
          $ref: '#/components/responses/TagResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/TagNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /tags/{id}:
    get:
      summary: Get the tag by id.
      operationId: getTagById
      tags:
        - Tags
      parameters:
        - name: id
          in: path
          description: Tag id.
          required: true
          example: 5f66a68227efbb73c2996c1a
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/TagResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/TagNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete the tag by id.
      operationId: deleteTagById
      tags:
        - Tags
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Tag id.
          required: true
          example: 5f66a68227efbb73c2996c1a
          schema:
            type: string
      responses:
        200:
          description: 200 OK
        400:
          $ref: '#/components/responses/BadRequestError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/TagNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'

  # Paths :: Comments
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    InternalServerError:
      description: Internal Server Error.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/InternalServerError'
    BadRequestError:
      description: Bad Request.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/BadRequestError'
    UnauthorizedError:
      description: Unauthorized.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
...
    TagResponse:
      description: Tag response.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/TagResponse'
    TagListResponse:
      description: List of tags.
      content:
        'application/json':
          schema:
            type: array
            items:
              $ref: '#/components/schemas/TagResponse'
    TagNotFoundError:
      description: Tag not found error.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/TagNotFoundError'
...
  schemas:
...
    InternalServerError:
      type: object
      properties:
        title:
          type: string
          description: A title of error response.
          example: Internal Server Error
        code:
          type: string
          description: A code of error response.
          example: RP-1000
        message:
          type: string
          description: A message of error response.
          example: Internal Server Error
    BadRequestError:
      type: object
      properties:
        title:
          type: string
          description: A title of error response.
          example: Bad Request
        code:
          type: string
          description: A code of error response.
          example: RP-1001
        message:
          type: string
          description: A message of error response.
          example: The message cannot be read.
    UnauthorizedError:
      type: object
      properties:
        title:
          type: string
          description: A title of error response.
          example: Unauthorized
        code:
          type: string
          description: A code of error response.
          example: RP-1006
        message:
          type: string
          description: A message of error response.
          example: Unauthorized.
...
    TagNotFoundError:
      type: object
      properties:
        title:
          type: string
          description: A title of error response.
          example: Tag Not Found Error
        code:
          type: string
          description: A code of error response.
          example: RP-1003 # FIXME: Please, fix it.
        message:
          type: string
          description: A message of error response.
          example: The tag by id '5f3193eee810fe1a026e7f84' is not found.
...
    TagRequest:
      type: object
      properties:
        id:
          type: string
          description: Tag id.
          example: 5f66a68227efbb73c2996c1a
        name:
          type: string
          description: Tag name.
          example: My first tag
...
    TagResponse:
      type: object
      properties:
        id:
          type: string
          description: Tag id.
          example: 3579626af2a44ee983a299fc65f7238f
        name:
          type: string
          description: Tag name.
          example: java

    # Schemas :: Enums
...